changes in Material, Item, what can fill the room
An item can be :
- a Light(not properly physical, does not absorb rays)
- a physicalObject (like before)
- a Lamp (basically a physical object which emits light)

A material is either opaque (OpaqueObject) or transparent
(LightConductor).
Transparent means it is a LightCoductor with a refractive index. Opaque
means that light cannot be transmitted through it.

A Transparent object conducts light through a medium (attribute of the
class).

Loader has been updated to use these classes.